connectors/docker/requirements.txt
requests==2.32.3
psutil==6.1.0
websocket-client==1.8.0
